Book_writer_Crew/pyproject.toml
[project]
name = "book-writer-crew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Khurramshahzad175", email = "khurramshahzad2440@gmail.com" }
]
requires-python = ">=3.12.8"
dependencies = [
    "crewai>=0.100.1",
    "crewai-tools>=0.33.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
book-writer-crew = "book_writer_crew:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
